requirements.txt
django == 5.0.8
djangorestframework == 3.15.2
mysqlclient == 2.2.4


asgiref              == 3.8.1
Django               == 5.0.8
djangorestframework  == 3.15.2
mysqlclient          == 2.2.4
pip                  == 24.2
sqlparse             == 0.5.1
tzdata               == 2024.1
